requirements.txt
anthropic>=0.39.0
google-genai>=1.0.0
requests>=2.31.0
pyyaml>=6.0.1
